requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
redis==4.6.0
openai==0.28.1
konlpy==0.6.0
requests==2.31.0
pytest==7.4.0
httpx==0.23.0
python-dotenv==1.0.0
bcrypt==4.0.1
google-cloud-translate==3.11.1
google-auth==2.23.4 